Overhauled export action to match other MeasurementViewerActions
