+++ components/insight/SRC/org/openmicroscopy/shoola/agents/measurement/view/MeasurementViewerUI.java
@@ -49,17 +49,19 @@ import javax.swing.AbstractButton;
 import javax.swing.ButtonGroup;
 import javax.swing.ImageIcon;
 import javax.swing.JCheckBoxMenuItem;
+import javax.swing.JFileChooser;
+import javax.swing.JFrame;
 import javax.swing.JMenu;
 import javax.swing.JMenuBar;
 import javax.swing.JMenuItem;
 import javax.swing.JTabbedPane;
+import javax.swing.SwingUtilities;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import javax.swing.filechooser.FileFilter;
 
 //Third-party libraries
 import org.jhotdraw.draw.AttributeKey;
-import org.jhotdraw.draw.AttributeKeys;
 import org.jhotdraw.draw.DelegationSelectionTool;
 import org.jhotdraw.draw.Drawing;
 import org.jhotdraw.draw.Figure;
@@ -80,6 +82,8 @@ import org.openmicroscopy.shoola.env.ui.TaskBar;
 import org.openmicroscopy.shoola.env.ui.TopWindow;
 import org.openmicroscopy.shoola.env.ui.UserNotifier;
 import org.openmicroscopy.shoola.util.filter.file.ExcelFilter;
+import org.openmicroscopy.shoola.util.filter.file.JPEGFilter;
+import org.openmicroscopy.shoola.util.filter.file.PNGFilter;
 import org.openmicroscopy.shoola.util.roi.exception.NoSuchROIException;
 import org.openmicroscopy.shoola.util.roi.exception.ROICreationException;
 import org.openmicroscopy.shoola.util.roi.model.annotation.MeasurementAttributes;
@@ -93,6 +97,7 @@ import org.openmicroscopy.shoola.util.ui.LoadingWindow;
 import org.openmicroscopy.shoola.util.ui.UIUtilities;
 import org.openmicroscopy.shoola.util.ui.drawingtools.canvas.DrawingCanvasView;
 import org.openmicroscopy.shoola.util.ui.filechooser.FileChooser;
+import org.openmicroscopy.shoola.util.ui.graphutils.ChartObject;
 
 /** 
  * The {@link MeasurementViewer} view.
@@ -1593,4 +1598,31 @@ class MeasurementViewerUI
 		super.setVisible(value);
 	}
 
+ 	/**
+         * Opens a file chooser dialog and exports the graph as JPEG or PNG.
+         */
+        public void exportGraph() {
+            List<FileFilter> filterList = new ArrayList<FileFilter>();
+            filterList.add(new JPEGFilter());
+            filterList.add(new PNGFilter());
+            FileChooser chooser = new FileChooser(
+                    (JFrame) SwingUtilities.windowForComponent(this),
+                    FileChooser.SAVE, "Save Graph",
+                    "Save the graph as JPEG or PNG", filterList);
+            try {
+                File f = UIUtilities.getDefaultFolder();
+                if (f != null)
+                    chooser.setCurrentDirectory(f);
+            } catch (Exception ex) {
+            }
+            if (chooser.showDialog() != JFileChooser.APPROVE_OPTION)
+                return;
+            File file = chooser.getFormattedSelectedFile();
+            FileFilter filter = chooser.getSelectedFilter();
+    
+            int type = (filter instanceof JPEGFilter) ? ChartObject.SAVE_AS_JPEG
+                    : ChartObject.SAVE_AS_PNG;
+    
+            graphPane.saveGraph(file, type);
+        }
 }
